tests/wpt/tests/html/canvas/offscreen/shadows/2d.shadow.attributes.shadowOffset.valid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.shadow.attributes.shadowOffset.valid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.shadow.attributes.shadowOffset.valid</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.shadowOffsetX = 1;
  ctx.shadowOffsetY = 2;
  _assertSame(ctx.shadowOffsetX, 1, "ctx.shadowOffsetX", "1");
  _assertSame(ctx.shadowOffsetY, 2, "ctx.shadowOffsetY", "2");

  ctx.shadowOffsetX = 0.5;
  ctx.shadowOffsetY = 0.25;
  _assertSame(ctx.shadowOffsetX, 0.5, "ctx.shadowOffsetX", "0.5");
  _assertSame(ctx.shadowOffsetY, 0.25, "ctx.shadowOffsetY", "0.25");

  ctx.shadowOffsetX = -0.5;
  ctx.shadowOffsetY = -0.25;
  _assertSame(ctx.shadowOffsetX, -0.5, "ctx.shadowOffsetX", "-0.5");
  _assertSame(ctx.shadowOffsetY, -0.25, "ctx.shadowOffsetY", "-0.25");

  ctx.shadowOffsetX = 0;
  ctx.shadowOffsetY = 0;
  _assertSame(ctx.shadowOffsetX, 0, "ctx.shadowOffsetX", "0");
  _assertSame(ctx.shadowOffsetY, 0, "ctx.shadowOffsetY", "0");

  ctx.shadowOffsetX = 1e6;
  ctx.shadowOffsetY = 1e6;
  _assertSame(ctx.shadowOffsetX, 1e6, "ctx.shadowOffsetX", "1e6");
  _assertSame(ctx.shadowOffsetY, 1e6, "ctx.shadowOffsetY", "1e6");
  t.done();

});
</script>
